+++ src/io/BpmnIO.js
@@ -1,6 +1,7 @@
 export default function BpmnIO(activity, context) {
   this.activity = activity;
   this.context = context;
+  this.type = 'bpmnio';
 
   const {
     ioSpecification: ioSpecificationDef,
@@ -10,6 +11,12 @@ export default function BpmnIO(activity, context) {
   this.specification = ioSpecificationDef && new ioSpecificationDef.Behaviour(activity, ioSpecificationDef, context);
   this.properties = propertiesDef && new propertiesDef.Behaviour(activity, propertiesDef, context);
 }
+
+Object.defineProperty(BpmnIO.prototype, 'hasIo', {
+  get() {
+    return this.specification || this.properties;
+  },
+});
 
 BpmnIO.prototype.activate = function activate(message) {
   const properties = this.properties, specification = this.specification;